requirements.txt
Django==1.8.3
django-widget-tweaks==1.4.1
django-zurb-foundation==5.5.0

# python manage.py graph_models --pygraphviz -a -g -o my_project_visualized.svg
django-extensions==1.5.5
pygraphviz==1.3rc2

# testing
factory_boy==2.5.2
